+++ spyglass-sample/src/main/java/com/linkedin/android/spyglass/sample/SamplesActivity.java
@@ -16,10 +16,11 @@ package com.linkedin.android.spyglass.sample;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.v7.app.ActionBarActivity;
+import android.support.v7.app.AppCompatActivity;
 import android.view.View;
 import android.widget.Button;
 
+import com.linkedin.android.spyglass.sample.samples.ColorfulMentions;
 import com.linkedin.android.spyglass.sample.samples.GridMentions;
 import com.linkedin.android.spyglass.sample.samples.MultiSourceMentions;
 import com.linkedin.android.spyglass.sample.samples.NetworkedMentions;
@@ -28,7 +29,7 @@ import com.linkedin.android.spyglass.sample.samples.SimpleMentions;
 /**
  * Main samples activity containing buttons to launch all the different samples.
  */
-public class SamplesActivity extends ActionBarActivity {
+public class SamplesActivity extends AppCompatActivity {
 
     @Override
     public void onCreate(Bundle savedInstanceState) {
@@ -40,6 +41,14 @@ public class SamplesActivity extends ActionBarActivity {
             @Override
             public void onClick(View v) {
                 startActivity(new Intent(SamplesActivity.this, SimpleMentions.class));
+            }
+        });
+
+        Button colorfulSample = (Button) findViewById(R.id.colorful_sample);
+        colorfulSample.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                startActivity(new Intent(SamplesActivity.this, ColorfulMentions.class));
             }
         });