+++ spyglass/src/test/java/com/linkedin/android/spyglass/ui/RichEditorViewTest.java
@@ -83,22 +83,22 @@ public class RichEditorViewTest {
         assertEquals(0, mRichEditor.getMentionSpans().size());
 
         Mentionable mention1 = new TestMention("Shoulong Li");
-        MentionSpan span1 = new MentionSpan(Robolectric.application, mention1);
+        MentionSpan span1 = new MentionSpan(mention1);
         Mentionable mention2 = new TestMention("Deepank Gupta");
-        MentionSpan span2 = new MentionSpan(Robolectric.application, mention2);
+        MentionSpan span2 = new MentionSpan(mention2);
 
-        mRichEditor.getText().setSpan(span1, 14, 14 + mention1.getPrimaryText().length(), Spanned.SPAN_EXCLUSIVE_EXCLUSIVE);
-        mRichEditor.getText().setSpan(span2, 30, 30 + mention2.getPrimaryText().length(), Spanned.SPAN_EXCLUSIVE_EXCLUSIVE);
+        mRichEditor.getText().setSpan(span1, 14, 14 + mention1.getSuggestiblePrimaryText().length(), Spanned.SPAN_EXCLUSIVE_EXCLUSIVE);
+        mRichEditor.getText().setSpan(span2, 30, 30 + mention2.getSuggestiblePrimaryText().length(), Spanned.SPAN_EXCLUSIVE_EXCLUSIVE);
         List<MentionSpan> spans = mRichEditor.getMentionSpans();
 
         assertEquals(2, spans.size());
         Mentionable result1 = spans.get(0).getMention();
         Mentionable result2 = spans.get(1).getMention();
 
-        assertEquals(mention1.getPrimaryText(), result1.getPrimaryText());
-        assertEquals(mention1.getPrimaryText(), "Shoulong Li");
-        assertEquals(mention2.getPrimaryText(), result2.getPrimaryText());
-        assertEquals(mention2.getPrimaryText(), "Deepank Gupta");
+        assertEquals(mention1.getSuggestiblePrimaryText(), result1.getSuggestiblePrimaryText());
+        assertEquals(mention1.getSuggestiblePrimaryText(), "Shoulong Li");
+        assertEquals(mention2.getSuggestiblePrimaryText(), result2.getSuggestiblePrimaryText());
+        assertEquals(mention2.getSuggestiblePrimaryText(), "Deepank Gupta");
     }
 
     @Test